+++ packages/outbox/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 0.9.37
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @volontariapp/database@3.4.0
+
 ## 0.9.36
 
 ### Patch Changes